test car specific stuff in their own files
