+++ selfdrive/car/subaru/tests/test_subaru.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python3
+import unittest
+
+from cereal import car
+from selfdrive.car.subaru.values import FW_VERSIONS
+
+Ecu = car.CarParams.Ecu
+ECU_NAME = {v: k for k, v in Ecu.schema.enumerants.items()}
+
+
+class TestChryslerFingerprint(unittest.TestCase):
+  def test_blacklisted_ecus(self):
+    blacklisted_addrs = (0x7c4, 0x7d0)  # includes A/C ecu and an unknown ecu
+    for car_model, ecus in FW_VERSIONS.items():
+      with self.subTest(car_model=car_model):
+        for ecu in ecus.keys():
+          self.assertNotIn(ecu[1], blacklisted_addrs, f'{car_model}: Blacklisted ecu: (Ecu.{ECU_NAME[ecu[0]]}, {hex(ecu[1])})')
+
+
+if __name__ == "__main__":
+  unittest.main()
